Migrate partially to bsp v2.0
Start the migration to bsp 2.0. There's not an official release yet but
we're going to depend on a milestone to ease the migration of the
clients and their testing via Bloop.

The following migration only misses support for `test`, `run` and a few
other endpoints that we will implement in a future PR.